requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
colorama==0.4.6
Django==4.0.4
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.6
